requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
WTForms>=3.1.0
email-validator>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
requests>=2.28.0
